requirements.txt
pytz
datetime

cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Session==0.3.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
multitasking==0.0.9
numpy==1.19.2
pip==20.2.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
setuptools==49.6.0
six==1.15.0
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.35.1

